requirements.txt
fastapi==0.115.6
starlette==0.41.3
mangum==0.19.0
PyJWT==2.10.1
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.3.7
langchain-core==0.3.25
langchain-openai==0.2.12
langchain-mongodb==0.3.0
langchain-text-splitters==0.3.2
langgraph==0.2.59
langgraph-checkpoint==2.0.9
langgraph-checkpoint-mongodb==0.1.0
langgraph-sdk==0.1.47
langsmith==0.1.143
openai==1.58.1
certifi==2024.6.2
bcrypt==4.2.1
pymongo==4.9.2
typing_extensions==4.12.2
aiohttp==3.10.5
anyio==4.4.0
idna==3.7
sniffio==1.3.1
pydantic_core==2.23.4
annotated-types==0.7.0
dnspython==2.7.0
jsonpatch==1.33
packaging==24.1
PyYAML==6.0.1
tenacity==8.5.0
tiktoken==0.8.0
motor==3.6.0
numpy==1.26.4
msgpack==1.1.0
httpx==0.27.0
orjson==3.10.11
httpcore==1.0.5
h11==0.14.0
requests==2.32.3
requests-toolbelt==1.0.0
charset-normalizer==3.3.2
urllib3==2.2.1
distro==1.9.0
jiter==0.7.1
tqdm==4.67.0
aiohappyeyeballs==2.4.0
aiosignal==1.3.1
attrs==23.2.0
frozenlist==1.4.1
multidict==6.1.0
yarl==1.11.1
trio==0.25.1
jsonpointer==3.0.0
regex==2024.11.6
outcome==1.3.0.post0
sortedcontainers==2.4.0